# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Bringing group travel plans to life",
        "title": "Toggo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/comments/{commentID}/reactions": {
            "get": {
                "description": "Retrieves aggregated emoji reactions for a comment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment-reactions"
                ],
                "summary": "Get reactions summary for a comment",
                "operationId": "getCommentReactionsSummary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "commentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommentReactionsSummaryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds an emoji reaction to a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment-reactions"
                ],
                "summary": "Add a reaction to a comment",
                "operationId": "addCommentReaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "commentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create reaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCommentReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CommentReaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes the current user's emoji reaction from a comment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment-reactions"
                ],
                "summary": "Remove a reaction from a comment",
                "operationId": "removeCommentReaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "commentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delete reaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeleteCommentReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/comments/{commentID}/reactions/{emoji}/users": {
            "get": {
                "description": "Retrieves users who reacted with a specific emoji on a comment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comment-reactions"
                ],
                "summary": "Get users who reacted with an emoji",
                "operationId": "getCommentReactionUsers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "commentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Emoji (URL-encoded)",
                        "name": "emoji",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CommentReactionUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.APIError": {
            "type": "object",
            "properties": {
                "message": {},
                "statusCode": {
                    "type": "integer"
                }
            }
        },
...
        "models.CommentReaction": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.CommentReactionSummary": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "emoji": {
                    "type": "string"
                },
                "reacted_by_me": {
                    "type": "boolean"
                }
            }
        },
        "models.CommentReactionUser": {
            "type": "object",
            "properties": {
                "profile_picture_url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.CommentReactionUsersResponse": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CommentReactionUser"
                    }
                }
            }
        },
        "models.CommentReactionsSummaryResponse": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "reactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CommentReactionSummary"
                    }
                }
            }
        },
...
        "models.CreateCommentReactionRequest": {
            "type": "object",
            "required": [
                "emoji"
            ],
            "properties": {
                "emoji": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "models.DeleteCommentReactionRequest": {
            "type": "object",
            "required": [
                "emoji"
            ],
            "properties": {
                "emoji": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
    }
}